little refactoring, fix small errors, fixing field types

+++ CarRental.Shared/Interfaces/IBooking.cs
@@ -10,7 +10,7 @@ public interface IBooking
     DateTime ReturnDate { get; set; }
     int VehicleId { get; }
     int CustomerId { get; }
-    decimal? TotalCost { get; }
+    double? TotalCost { get; }
 
     void CalculateTotalCost(Vehicle vehicle, int kmDriven);
 }
